pizzapoint/requirements.txt
django
djangorestframework
pillow
django-cors-headers
twilio
djangorestframework-simplejwt